backend/requirements.txt
# BlueWall Backend Dependencies
# Core FastAPI framework and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0

# Enhanced Security and Cryptography
cryptography==41.0.7
pycryptodome==3.19.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Database Drivers (uncomment as needed)
# PostgreSQL
# psycopg2-binary==2.9.9

# MySQL
# pymysql==1.1.0

# SQLite (included with Python)

# Utilities
python-multipart==0.0.6
email-validator==2.1.0

# Development and Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and Monitoring
structlog==23.2.0

# CORS and Middleware
fastapi-cors==0.0.6

# WebSocket Support
websockets==12.0

# Rate Limiting (optional)
slowapi==0.1.9

# Health Checks
healthcheck==1.3.0

# Enhanced Security Monitoring
psutil>=5.9.0
pywin32>=306; sys_platform == "win32"
pyobjc-framework-Security>=9.2; sys_platform == "darwin"
pyobjc-framework-SystemConfiguration>=9.2; sys_platform == "darwin"

# Memory and Process Analysis
memory-profiler>=0.61.0

# Production Server (alternative to uvicorn)
# gunicorn==21.2.0

# Note: Install with: pip install -r requirements.txt
# For production, consider using a virtual environment:
# python -m venv venv
# source venv/bin/activate  # On Windows: venv\Scripts\activate
# pip install -r requirements.txt
